Correct wrong typehint

+++ ex2/ft_plant_growth.py
@@ -7,7 +7,7 @@ Demonstrates creating and displaying multiple plants.
 class Plant:
     """Represents a plant with basic growth metrics."""
 
-    def __init__(self, name: str, height: int = 0, age: int = 0) -> None:
+    def __init__(self, name: str, height: float = 0, age: int = 0) -> None:
         """Initialize a plant with name, height (cm), and age (days)."""
         if height >= 0 and age >= 0 and name != "":
             self.name = name
